% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_asymp.r
\name{estimator_asymp}
\alias{estimator_asymp}
\title{asymptotic estimator for richness}
\usage{
estimator_asymp(x, ...)
}
\arguments{
\item{x}{A vector of species IDs.}

\item{...}{Other options for iNEXT::iNEXT().}
}
\value{
A data frame with estimated asymptote of species richness, SE, and 95% CI.
}
\description{
Get the asymptotic estimator for richness, and the bounds of its 95% conf int
}
