% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_chao1.r
\name{estimator_chao1}
\alias{estimator_chao1}
\title{Chao1 richness estimator}
\usage{
estimator_chao1(x)
}
\arguments{
\item{x}{A vector of species IDs.}
}
\value{
A data frame of Chao 1 estiamtion and variance and 95% CI.
}
\description{
Chao1 richness estimator
}
