% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_google_drive.r
\name{read_from_google_drive}
\alias{read_from_google_drive}
\title{read from google drive}
\usage{
read_from_google_drive(file_name_string = NULL,
  my_path_to_googledirve_directory = NULL,
  keep_local_copy_of_file = TRUE, ...)
}
\arguments{
\item{file_name_string}{Data frame you want to write out (dataframe or tibble).}

\item{my_path_to_googledirve_directory}{Path in Google Drive directory (google id object, or text string).}

\item{keep_local_copy_of_file}{Do you want to keep file locally (boolean, default = TRUE)?}

\item{...}{other arguments passed to readr::read_csv}
}
\description{
read a csv or txt file from google drive
}
\details{
Read file from Google Drive
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Eric R. Sokol \email{esokol@battelleecology.org}
}
