% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_cumulative.r
\name{richness_cumulative}
\alias{richness_cumulative}
\title{cumulative richness estimators by site and year}
\usage{
richness_cumulative(dat, sp_name = "taxonID", site_name = "siteID",
  plot_name = "plotID", value_name = "value", year_name = "year",
  grp_by = c("site", "site_plot"), cumul_by_year = TRUE)
}
\arguments{
\item{dat}{The name of the data frame.}

\item{sp_name}{Column name of species/taxon, a single string.}

\item{site_name}{Column name of site, a single string.}

\item{plot_name}{Column name of plot (if any), a single string.}

\item{value_name}{Column name of abundance/cover/density/presence/absence, a single string.}

\item{year_name}{Column name of year, a single string.}

\item{grp_by}{Whether to group by site or group by both site and group.}

\item{cumul_by_year}{To calculate richness culumatively through sample years? Default is `TRUE`.}
}
\value{
A data frame of richness, Chao1 etimates, asymptote estimates and their uncertainties.
}
\description{
Function to get cumulative richness estimators by site and year. Sequentially add
years and see what happens to the cumulative observed richness and richness estimators. Each
year's result represents all data up to that point in time.
}
