% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_google_drive.r
\name{write_to_google_drive}
\alias{write_to_google_drive}
\title{write to google drive}
\usage{
write_to_google_drive(data_to_write, write_filename = NULL,
  my_path_to_googledirve_directory, keep_local_copy_of_file = TRUE)
}
\arguments{
\item{data_to_write}{Data frame you want to write out (dataframe or tibble).}

\item{write_filename}{File name to be saved (character).}

\item{my_path_to_googledirve_directory}{Path in Google Drive directory (google id object, or text string).}

\item{keep_local_copy_of_file}{Do you want to keep file locally (boolean)?}

\item{...}{Other arguments passed to readr::write_csv}
}
\description{
writes to google drive
}
\details{
Write file to Google Drive
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Eric R. Sokol \email{esokol@battelleecology.org}
}
